The rest of the report card: one vocabulary, one host resolver, a net guard, a mode machine, a lean page, and no more e2e sleeps
Effect/overlay/palette ids live in tools/effect_vocab.py and a test reads
effects.ts and castle_effects.h to prove all three agree; hosts.candidates
is the single resolver castle_link uses; background jobs queue behind the
same lock as sync encodes; import arguments can no longer become flags;
URL imports refuse private/loopback targets unless you are on the Mac;
multipart names '.'/'..' are 400s; the C++ harness fails loudly in CI
when no compiler exists; the suite prints nothing but OK; mypy checks
untyped bodies and unused ignores. The desk's six audition flags are one
typed mode machine; panels.selectScene replaces selector reach-arounds;
api.ts is the only doorway; the studio serves a lean page that streams
scene audio instead of inlining 1.9 MB. docs/PARITY.md and CONTRIBUTING
exist; the e2e suite has zero timed sleeps and lane-derived ports.
ESPHome 2026.8.0 was trialled in isolation and still overflows dram0 by
16 bytes — held, documented.

+++ tests/test_analysis.py
@@ -239,7 +239,7 @@ class TestDensityFitting(unittest.TestCase):
 
     def residual(self, decay: float, gap: float) -> float:
         """How much of a flash survives to the next hit."""
-        return decay ** (gap / it.FRAME)
+        return float(decay ** (gap / it.FRAME))
 
     def test_dense_material_decays_faster(self) -> None:
         decay, _ = it.fit_to_density(self.hits(0.24), 0.92)
